+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+* improve **FileSourceProcessor:** improve parsing error logging
+
 ## [0.10.0] - 2022-07-19
 
 ### Added